requirements.txt
google-genai>=1.5.0
chromadb>=0.4.24
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
